backend/pyproject.toml
[tool.poetry]
name = "education-platform"
version = "0.1.0"
description = ""
authors = ["Dmitry Seregin <dmitry.seregin.msk@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = []
package-mode = false

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.6"
pre-commit = "^3.1.1"


[tool.poetry.dependencies]
python = "^3.11"
anyio = "^3.6.2"
click = "^8.1.3"
DateTime = "^5.1"
fastapi = "^0.101.1"
h11 = "^0.14.0"
idna = "^3.4"
pydantic = "^2.1.1"
pytz = "^2023.3"
sniffio = "^1.3.0"
SQLAlchemy = "^2.0.10"
starlette = "^0.27.0"
typing-extensions = "^4.5.0"
uvicorn = "^0.21.1"
websockets = "^10.4"
"zope.interface" = "^6.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
fastapi-pagination = "^0.12.8"
celery = "^5.2.7"
redis = "^4.5.4"
requests = "^2.28.2"
boto3 = "^1.26.102"
fastapi-mail = "^1.2.6"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.6"
starlette-admin = {extras = ["i18n"], version = "^0.10.0"}
structlog = "^23.1.0"
alembic = "^1.11.1"
itsdangerous = "^2.1.2"
bump-pydantic = "^0.6.1"
sentry-sdk = {extras = ["fastapi"], version = "^1.29.2"}
prometheus-fastapi-instrumentator = "^6.1.0"
asyncpg = "^0.29.0"
httpx = "^0.27.0"
greenlet = "^3.0.3"
bcrypt = "^4.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
